# Conflicts:
#	database/src/main/resources/Mybatis-Generator.xml

+++ service/src/main/java/cn/thinkfree/service/approvalflow/impl/ApprovalFlowNodeRoleServiceImpl.java
@@ -45,8 +45,10 @@ public class ApprovalFlowNodeRoleServiceImpl implements ApprovalFlowNodeRoleServ
 
     @Override
     public void deleteByNodeNums(List<String> nodeNums) {
-        ApprovalFlowNodeRoleExample example = new ApprovalFlowNodeRoleExample();
-        example.createCriteria().andNodeNumIn(nodeNums);
-        nodeRoleMapper.deleteByExample(example);
+        if (nodeNums != null && nodeNums.size() > 0 ) {
+            ApprovalFlowNodeRoleExample example = new ApprovalFlowNodeRoleExample();
+            example.createCriteria().andNodeNumIn(nodeNums);
+            nodeRoleMapper.deleteByExample(example);
+        }
     }
 }
